updates
made major updates to the website -- do not push this live yet (need to address a few more pages). The highscores should be ready to start filling with data. There is the main highscores, as well as a page dedicated to display each user's data (click on the "username" on the highscores).